Refactor the Assembly module Add NCBISequence class that can download single sequences
