+++ ebi_eva_common_pyutils/reference/__init__.py
@@ -1,0 +1,1 @@
+from ebi_eva_common_pyutils.reference.assembly import NCBIAssembly
